requirements.txt
numpy
tqdm
pandas
biopython